wip: way better messages UI
